mlv_rec: fixed chroma smooth regression in mlv_dump caused by API change in 5167619. thanks dfort/danne/bouncyball.

+++ modules/mlv_rec/mlv_dump.c
@@ -928,8 +928,8 @@ void chroma_smooth(int method, struct raw_info *info)
     int w = info->width;
     int h = info->height;
 
-    unsigned short * aux = malloc(w * h * sizeof(short));
-    unsigned short * aux2 = malloc(w * h * sizeof(short));
+    uint32_t * aux = malloc(w * h * sizeof(uint32_t));
+    uint32_t * aux2 = malloc(w * h * sizeof(uint32_t));
 
     int x,y;
     for (y = 0; y < h; y++)
